[add]Steam連携のときに強調表示するように変更
