+++ src/app/page.tsx
@@ -1,26 +1,16 @@
+import { Providers } from "@/components/common/AuthProvider";
 import Network from "@/components/network/Network";
-import { SteamDetailsDataType } from "@/types/api/getSteamDetailType";
 import { SteamListType } from "@/types/NetworkType";
 
 export default async function Page() {
-  try {
-    // データ取得
-    const steamListRes = await fetch(
-      `${process.env.NEXT_PUBLIC_CURRENT_URL}/api/network/getSteamList`
-    );
-    const steamListData: SteamListType[] = await steamListRes.json();
+  const steamListRes = await fetch(
+    `${process.env.NEXT_PUBLIC_CURRENT_URL}/api/network/getSteamList`
+  );
+  const steamListData: SteamListType[] = await steamListRes.json();
 
-    return (
-      <div className="flex flex-col h-screen">
-        <Network steamListData={steamListData} />
-      </div>
-    );
-  } catch (error) {
-    console.error("Error fetching data:", error);
-    return (
-      <div className="flex flex-col h-screen">
-        <p>データの取得に失敗しました。</p>
-      </div>
-    );
-  }
+  return (
+    <div className="flex flex-col h-screen">
+      <Providers><Network steamListData={steamListData} /></Providers>
+    </div>
+  );
 }
